+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## [UNRELEASED] go-datareader 0.2.0
 
 ### Features
-- Add `Tiingo`` data reader ([#2](https://github.com/AleksanderWWW/go-datareader/pull/2))
+- Add `Tiingo` daily data reader ([#2](https://github.com/AleksanderWWW/go-datareader/pull/2))
 
 ### Changes
 - Repo cleanup ([#1](https://github.com/AleksanderWWW/go-datareader/pull/1))